mujoco-ur10e-auto-drilling/requirements.txt
mujoco>=3.1.0
numpy>=1.24.0
PyYAML>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
